MDL-61324 enrol_lti: Fix repeatedly sent unchanged grades
When invoking moodle lti grade sync, unchanged grades are repeatedly
sent on subsequent syncs to the consumer system regardless of whether
they have changed, if the `grade` variable is a float.

The comparison of `grade` and `lastgrade` will fail when `grade` is a
float, because `lastgrade` is cast to an int when saved to the
`enrol_lti_users` database table.

This fix removes the casting to an int when updating `lastgrade`.